# server/src/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Edtech - Sistema de Gestão Acadêmica",
        "description": "API para sistema de gestão acadêmica da instituição Edtech, permitindo autenticação e gerenciamento de matrículas de alunos em turmas online.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/students": {
            "get": {
                "summary": "Listar estudantes",
                "description": "Retorna uma lista paginada de estudantes. **Requer autenticação JWT.**",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número de itens por página",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100,
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de estudantes retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Estudantes listados com sucesso"
                                                },
                                                "students": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "João Silva"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "joao@email.com"
                                                            },
                                                            "academicRecord": {
                                                                "type": "string",
                                                                "example": "2023001"
                                                            },
                                                            "cpf": {
                                                                "type": "string",
                                                                "example": "12345678901"
                                                            },
                                                            "createdAt": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2023-01-15T10:30:00.000Z"
                                                            },
                                                            "updatedAt": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2023-01-15T10:30:00.000Z"
                                                            }
                                                        }
                                                    }
                                                },
                                                "pagination": {
                                                    "type": "object",
                                                    "properties": {
                                                        "page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "limit": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 25
                                                        },
                                                        "totalPages": {
                                                            "type": "integer",
                                                            "example": 3
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token de acesso requerido ou inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "auth": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "No token provided."
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação nos parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bad request"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "error": {
                                                    "type": "object",
                                                    "properties": {
                                                        "message": {
                                                            "type": "string",
                                                            "example": "Parâmetros de paginação inválidos"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Criar novo estudante",
                "description": "Cria um novo estudante no sistema. **Requer autenticação JWT.**",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@email.com"
                                    },
                                    "academicRecord": {
                                        "type": "string",
                                        "example": "2023001"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678901"
                                    }
                                },
                                "required": [
                                    "name",
                                    "email",
                                    "academicRecord",
                                    "cpf"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Estudante criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Created"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Estudante criado com sucesso"
                                                },
                                                "student": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "João Silva"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "joao@email.com"
                                                        },
                                                        "academicRecord": {
                                                            "type": "string",
                                                            "example": "2023001"
                                                        },
                                                        "cpf": {
                                                            "type": "string",
                                                            "example": "12345678901"
                                                        },
                                                        "createdAt": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2023-01-15T10:30:00.000Z"
                                                        },
                                                        "updatedAt": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2023-01-15T10:30:00.000Z"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token de acesso requerido ou inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "auth": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "No token provided."
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação nos dados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bad request"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "error": {
                                                    "type": "object",
                                                    "properties": {
                                                        "message": {
                                                            "type": "string",
                                                            "example": "Email já está em uso"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}